+++ changelog.md
@@ -7,7 +7,9 @@
  * add cdb pairs to output when no format is specified
  * when cdb or abuseipdb are present, without a format, all key pairs will be
    shown
- * option makethymecdb to download IP information from thyme.apnic.net
+ * option makethymecdb to download IP information from thyme.apnic.net, options
+   --data-raw-table, --ipv6-raw-table, --data-used-autnums can be used with a
+   file or URL
 
 0.2.7